gs_sw/requirements.txt
orjson
msgspec
websockets
pyserial-asyncio-fast